+++ firmware/application/src/data_cmd.h
@@ -68,6 +68,7 @@
 #define DATA_CMD_HF14A_RAW                      (2010)
 #define DATA_CMD_MF1_MANIPULATE_VALUE_BLOCK     (2011)
 #define DATA_CMD_MF1_CHECK_KEYS_OF_SECTORS      (2012)
+#define DATA_CMD_MF1_HARDNESTED_ACQUIRE         (2013)
 
 //
 // ******************************************************************